Added Security

+++ controllers/order.controller.ts
@@ -2,6 +2,7 @@ import { Request, Response, NextFunction } from 'express';
 import OrderService from '../services/order.service';
 import { MongoDbOrderRepository } from '../repositories/Order/mongodb.order.repository';
 import { MongoDbProductRepository } from '../repositories/Product/mongodb.product.repository';
+import AppError from '../utils/AppError';
 
 class OrderController {
   private orderService: OrderService;
@@ -36,6 +37,11 @@ class OrderController {
     try {
       const { id } = req.params;
       const { status } = req.body;
+      
+      if (req.role !== 'admin') {
+        throw new AppError('Unauthorized', 403);
+      }
+
       const updatedOrder = await this.orderService.updateOrderStatus(id, status);
       res.status(200).json(updatedOrder);
     } catch (error) {
@@ -46,6 +52,11 @@ class OrderController {
   public deleteOrder = async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { id } = req.params;
+
+      if (req.role !== 'admin') {
+        throw new AppError('Unauthorized', 403);
+      }
+
       const deletedOrder = await this.orderService.deleteOrder(id);
       res.status(200).json(deletedOrder);
     } catch (error) {
